second commit

+++ function-3-5.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-double sum_even(int array[], int n) {
+double sum_even(double array[], int n) {
     if(n<1){
         return 0;
     }else{
