+++ main-3-5.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 #include <stdlib.h>
 
-extern double sum_even(int*,int);
+extern double sum_even(double*,int);
 
 int main(int argc,char **argv)
 {
-    int array[5] = {1,2,3,4,5};
+    double array[5] = {1.0,2.0,3.0,4.0,5.0};
 	std :: cout << sum_even(array,5) << std :: endl;
 	return 0 ;
 }
